requirements.txt
Flask
psycopg2-binary
mysql-connector-python
APScheduler
click
pytest